+++ package.json
@@ -4,8 +4,8 @@
   "description": "This is an Express service that provides authorization functionality and includes gen-AI features.",
   "main": "index.js",
   "scripts": {
-    "start": "node src/app.js",
-    "dev": "nodemon src/app.js"
+    "start": "npx nodemon",
+    "dev": "tsx watch src/app.ts"
   },
   "keywords": [
     "AI",
@@ -17,14 +17,22 @@
   "author": "William Amaral",
   "license": "ISC",
   "dependencies": {
+    "@langchain/core": "^0.3.3",
     "@langchain/google-genai": "^0.1.0",
     "dotenv": "^16.4.5",
     "express": "^4.21.0",
-    "langchain": "^0.2.19",
+    "langchain": "^0.3.2",
     "multer": "1.4.5-lts.1"
   },
   "devDependencies": {
+    "@types/cors": "^2.8.17",
+    "@types/express": "^4.17.21",
+    "@types/multer": "^1.4.12",
+    "@types/node": "^22.5.5",
     "cors": "^2.8.5",
-    "nodemon": "^3.1.4"
+    "nodemon": "^3.1.4",
+    "ts-node": "^10.9.2",
+    "tsx": "^4.19.1",
+    "typescript": "^5.6.2"
   }
 }
