--- src/modules/gen-ai/routes.js
@@ -1,7 +1,0 @@
-const GenAIController = require('./GenAIController')
-const router = require('express').Router()
-
-router.post('/translate', GenAIController.translateText)
-router.post('/searchInDocument', GenAIController.searchInDocument)
-
-module.exports = router
