--- src/modules/status/routes.js
@@ -1,7 +1,0 @@
-const router = require('express').Router()
-
-router.get('/', (_, res) => {
-	res.status(200).send()
-})
-
-module.exports = router
